refactoring

+++ apps/currency/routers/schemas.py
@@ -1,5 +1,7 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class ShowCurrency(BaseModel):
-    result: float
+    result: Decimal
